Refactored SHP/Vxl handling by doing the same thing as FA2sp, and resolved shp turret display issues .

+++ 3rdParty/xcc/misc/cc_file.cpp
@@ -257,7 +257,7 @@ int Ccc_file::read()
 	return 0;
 }
 
-int Ccc_file::read(void* data, int size)
+int Ccc_file::read(void* data, int64_t size)
 {
 	if (get_p() < 0 || get_p() + size > get_size())
 		return 1;
